rename: common에서 shared로 디렉토리명 변경 2
