+++ frontend/src/shared/components/WaitingStatus.vue
